+++ festival-core/src/main/java/cn/bdqfork/core/context/ApplicationContext.java
@@ -1,0 +1,14 @@
+package cn.bdqfork.core.context;
+
+import cn.bdqfork.core.factory.BeanFactory;
+import cn.bdqfork.core.factory.ConfigurableBeanFactory;
+
+/**
+ * @author bdq
+ * @since 2020/1/8
+ */
+public interface ApplicationContext extends BeanFactory {
+    ConfigurableBeanFactory getConfigurableBeanFactory();
+
+    void close();
+}
