+++ festival-core/src/main/java/cn/bdqfork/core/exception/ResolvedException.java
@@ -4,7 +4,7 @@ package cn.bdqfork.core.exception;
  * @author bdq
  * @since 2019-07-30
  */
-public class ResolvedException extends Exception {
+public class ResolvedException extends BeansException {
     public ResolvedException(String message) {
         super(message);
     }
