+++ festival-core/src/main/java/cn/bdqfork/core/factory/ConfigurableBeanFactory.java
@@ -1,5 +1,7 @@
 package cn.bdqfork.core.factory;
 
+import cn.bdqfork.core.factory.processor.BeanPostProcessor;
+
 /**
  * @author bdq
  * @since 2019/12/15
