+++ festival-core/src/main/java/cn/bdqfork/core/factory/DefaultBeanFactory.java
@@ -1,6 +1,7 @@
 package cn.bdqfork.core.factory;
 
 import cn.bdqfork.core.exception.*;
+import cn.bdqfork.core.factory.definition.BeanDefinition;
 import cn.bdqfork.core.factory.registry.BeanDefinitionRegistry;
 import cn.bdqfork.core.util.BeanUtils;
 import cn.bdqfork.core.util.ReflectUtils;
@@ -133,7 +134,7 @@ public class DefaultBeanFactory extends AbstractAutoInjectedBeanFactory {
     }
 
     @Override
-    protected boolean containBeanDefinition(String beanName) {
+    public boolean containBeanDefinition(String beanName) {
         BeanFactory beanFactory = getParentBeanFactory();
         if (beanFactory instanceof AbstractBeanFactory) {
             AbstractBeanFactory abstractBeanFactory = (AbstractBeanFactory) beanFactory;
@@ -174,6 +175,15 @@ public class DefaultBeanFactory extends AbstractAutoInjectedBeanFactory {
     @Override
     public Map<String, BeanDefinition> getBeanDefinitions() {
         return beanDefinitionMap;
+    }
+
+    @Override
+    public void clearAllBeanDefinitions() {
+        beanDefinitionMap.clear();
+        if (getParentBeanFactory() instanceof BeanDefinitionRegistry) {
+            BeanDefinitionRegistry registry = (BeanDefinitionRegistry) getParentBeanFactory();
+            registry.clearAllBeanDefinitions();
+        }
     }
 
     @Override
