+++ festival-core/src/main/java/cn/bdqfork/core/factory/processor/BeanPostProcessor.java
@@ -1,4 +1,4 @@
-package cn.bdqfork.core.factory;
+package cn.bdqfork.core.factory.processor;
 
 import cn.bdqfork.core.exception.BeansException;
